test: break out into various test files
