+++ .circleci/config.yml
@@ -98,7 +98,7 @@ jobs:
   e2e-test:
     working_directory: ~/repo
     executor: nodeimage
-    parallelism: 7
+    parallelism: 11
     steps:
       - attach_workspace:
           at: ./
