--- test/integration/tests/veBalHelper.ts
@@ -1,552 +1,0 @@
-import { VeBalHelper } from '@custom-types/contracts';
-import { NamedAddresses, NamedContracts } from '@custom-types/types';
-import { SignerWithAddress } from '@nomiclabs/hardhat-ethers/signers';
-import { ProposalsConfig } from '@protocol/proposalsConfig';
-import { getAddresses, getImpersonatedSigner, time } from '@test/helpers';
-import { TestEndtoEndCoordinator } from '@test/integration/setup';
-import chai, { expect } from 'chai';
-import CBN from 'chai-bn';
-import { solidity } from 'ethereum-waffle';
-import { BigNumberish, Contract, Signer } from 'ethers';
-import { ethers } from 'hardhat';
-import { forceEth } from '../setup/utils';
-
-const e18 = (x: BigNumberish) => ethers.constants.WeiPerEther.mul(x);
-
-before(async () => {
-  chai.use(CBN(ethers.BigNumber));
-  chai.use(solidity);
-});
-
-describe('e2e-veBalHelper', function () {
-  const impersonatedSigners: { [key: string]: Signer } = {};
-  let contracts: NamedContracts;
-  let contractAddresses: NamedAddresses;
-  let deployAddress: string;
-  let e2eCoord: TestEndtoEndCoordinator;
-  let doLogging: boolean;
-  let vebalOtcHelper: Contract;
-  let otcBuyerAddress: string;
-  let otcBuyerSigner: SignerWithAddress;
-
-  let balWethBPTWhaleSigner: SignerWithAddress;
-  let gaugeTokenHolderSigner: SignerWithAddress;
-  const balWethBPTWhale = '0xC128a9954e6c874eA3d62ce62B468bA073093F25';
-  const balFeiWethGaugeTokenHolder = '0x7818a1da7bd1e64c199029e86ba244a9798eee10';
-
-  before(async function () {
-    deployAddress = (await ethers.getSigners())[0].address;
-    if (!deployAddress) throw new Error(`No deploy address!`);
-    const addresses = await getAddresses();
-    // add any addresses you want to impersonate here
-    const impersonatedAddresses = [addresses.userAddress, addresses.pcvControllerAddress, addresses.governorAddress];
-
-    doLogging = Boolean(process.env.LOGGING);
-
-    const config = {
-      logging: doLogging,
-      deployAddress: deployAddress,
-      version: 1
-    };
-
-    e2eCoord = new TestEndtoEndCoordinator(config, ProposalsConfig);
-
-    doLogging && console.log(`Loading environment...`);
-    ({ contracts, contractAddresses } = await e2eCoord.loadEnvironment());
-
-    ({ vebalOtcHelper } = contracts);
-    doLogging && console.log(`Environment loaded.`);
-    vebalOtcHelper = vebalOtcHelper as VeBalHelper;
-
-    for (const address of impersonatedAddresses) {
-      impersonatedSigners[address] = await getImpersonatedSigner(address);
-    }
-
-    otcBuyerAddress = contractAddresses.aaveCompaniesMultisig;
-    otcBuyerSigner = await getImpersonatedSigner(otcBuyerAddress);
-
-    balWethBPTWhaleSigner = await getImpersonatedSigner(balWethBPTWhale);
-    await forceEth(balWethBPTWhale);
-
-    await forceEth(balFeiWethGaugeTokenHolder);
-    gaugeTokenHolderSigner = await getImpersonatedSigner(balFeiWethGaugeTokenHolder);
-  });
-
-  it('should have correct owner and initial state', async () => {
-    expect(await vebalOtcHelper.owner()).to.equal(contractAddresses.aaveCompaniesMultisig);
-    expect(await vebalOtcHelper.pcvDeposit()).to.equal(contractAddresses.veBalDelegatorPCVDeposit);
-    expect(await vebalOtcHelper.balancerStaker()).to.equal(contractAddresses.balancerGaugeStaker);
-  });
-
-  it('onlyOwner should allow owner to call priviledged function', async () => {
-    await vebalOtcHelper.connect(otcBuyerSigner).setSpaceId(ethers.constants.HashZero);
-    expect(await contracts.veBalDelegatorPCVDeposit.spaceId()).to.be.eq(ethers.constants.HashZero);
-  });
-
-  it('all priviledged functions protected by onlyOwner should reject non-owner address calls', async () => {
-    const fakeOwner = ethers.constants.AddressZero;
-    const fakeOwnerSigner = await getImpersonatedSigner(fakeOwner);
-
-    const ADDRESS_ONE = '0x0000000000000000000000000000000000000001';
-
-    // setSpaceId
-    await expect(
-      contracts.vebalOtcHelper.connect(fakeOwnerSigner).setSpaceId(ethers.constants.HashZero)
-    ).to.be.revertedWith('Ownable: caller is not the owner');
-
-    // setDelegate
-    await expect(contracts.vebalOtcHelper.connect(fakeOwnerSigner).setDelegate(ADDRESS_ONE)).to.be.revertedWith(
-      'Ownable: caller is not the owner'
-    );
-
-    // clearDelegate
-    await expect(contracts.vebalOtcHelper.connect(fakeOwnerSigner).clearDelegate()).to.be.revertedWith(
-      'Ownable: caller is not the owner'
-    );
-
-    // setLockDuration
-    await expect(contracts.vebalOtcHelper.connect(fakeOwnerSigner).setLockDuration(5)).to.be.revertedWith(
-      'Ownable: caller is not the owner'
-    );
-
-    // lock
-    await expect(contracts.vebalOtcHelper.connect(fakeOwnerSigner).lock()).to.be.revertedWith(
-      'Ownable: caller is not the owner'
-    );
-
-    // exitLock
-    await expect(contracts.vebalOtcHelper.connect(fakeOwnerSigner).exitLock()).to.be.revertedWith(
-      'Ownable: caller is not the owner'
-    );
-
-    // setGaugeController
-    await expect(contracts.vebalOtcHelper.connect(fakeOwnerSigner).setGaugeController(ADDRESS_ONE)).to.be.revertedWith(
-      'Ownable: caller is not the owner'
-    );
-
-    // voteForGaugeWeight
-    await expect(
-      contracts.vebalOtcHelper
-        .connect(fakeOwnerSigner)
-        .voteForGaugeWeight(contractAddresses.dai, contractAddresses.weth, 5)
-    ).to.be.revertedWith('Ownable: caller is not the owner');
-
-    // stakeInGauge
-    await expect(
-      contracts.vebalOtcHelper.connect(fakeOwnerSigner).stakeInGauge(contractAddresses.dai, 5)
-    ).to.be.revertedWith('Ownable: caller is not the owner');
-
-    // stakeAllInGauge
-    await expect(
-      contracts.vebalOtcHelper.connect(fakeOwnerSigner).stakeAllInGauge(contractAddresses.dai)
-    ).to.be.revertedWith('Ownable: caller is not the owner');
-
-    // unstakeFromGauge
-    await expect(
-      contracts.vebalOtcHelper.connect(fakeOwnerSigner).unstakeFromGauge(contractAddresses.dai, 5)
-    ).to.be.revertedWith('Ownable: caller is not the owner');
-
-    // setBalancerMinter
-    await expect(
-      contracts.vebalOtcHelper.connect(fakeOwnerSigner).setBalancerMinter(contractAddresses.dai)
-    ).to.be.revertedWith('Ownable: caller is not the owner');
-
-    // transferBalancerStakerOwnership
-    await expect(
-      contracts.vebalOtcHelper.connect(fakeOwnerSigner).transferBalancerStakerOwnership(contractAddresses.dai)
-    ).to.be.revertedWith('Ownable: caller is not the owner');
-
-    // setVotingEscrowDelegation
-    await expect(
-      contracts.vebalOtcHelper.connect(fakeOwnerSigner).setVotingEscrowDelegation(contractAddresses.dai)
-    ).to.be.revertedWith('Ownable: caller is not the owner');
-    console.log('creating boost');
-
-    // create_boost
-    await expect(
-      contracts.vebalOtcHelper
-        .connect(fakeOwnerSigner)
-        .create_boost(contractAddresses.dai, contractAddresses.weth, '500', '10', '20', '0')
-    ).to.be.revertedWith('Ownable: caller is not the owner');
-
-    // extend_boost
-    await expect(
-      contracts.vebalOtcHelper.connect(fakeOwnerSigner).extend_boost('0', '500', '10', '20')
-    ).to.be.revertedWith('Ownable: caller is not the owner');
-
-    // cancel_boost
-    await expect(contracts.vebalOtcHelper.connect(fakeOwnerSigner).cancel_boost('0')).to.be.revertedWith(
-      'Ownable: caller is not the owner'
-    );
-
-    // burn
-    await expect(contracts.vebalOtcHelper.connect(fakeOwnerSigner).burn('0')).to.be.revertedWith(
-      'Ownable: caller is not the owner'
-    );
-
-    // withdrawERC20
-    await expect(
-      contracts.vebalOtcHelper.connect(fakeOwnerSigner).withdrawERC20(contractAddresses.dai, contractAddresses.weth, 5)
-    ).to.be.revertedWith('Ownable: caller is not the owner');
-
-    // withdrawETH
-    await expect(
-      contracts.vebalOtcHelper.connect(fakeOwnerSigner).withdrawETH(contractAddresses.dai, 5)
-    ).to.be.revertedWith('Ownable: caller is not the owner');
-
-    // withdrawERC20fromStaker
-    await expect(
-      contracts.vebalOtcHelper
-        .connect(fakeOwnerSigner)
-        .withdrawERC20fromStaker(contractAddresses.dai, contractAddresses.weth, 5)
-    ).to.be.revertedWith('Ownable: caller is not the owner');
-
-    // withdrawETHfromStaker
-    await expect(
-      contracts.vebalOtcHelper.connect(fakeOwnerSigner).withdrawETHfromStaker(contractAddresses.dai, 5)
-    ).to.be.revertedWith('Ownable: caller is not the owner');
-  });
-
-  describe('Setters', () => {
-    it('should be able to setSpaceId() to change the snapshot space id', async () => {
-      const beforeSnapshotId = await contracts.veBalDelegatorPCVDeposit.spaceId();
-      await vebalOtcHelper.connect(otcBuyerSigner).setSpaceId(ethers.constants.HashZero);
-      expect(await contracts.veBalDelegatorPCVDeposit.spaceId()).to.be.eq(ethers.constants.HashZero);
-      await vebalOtcHelper.connect(otcBuyerSigner).setSpaceId(beforeSnapshotId);
-      expect(await contracts.veBalDelegatorPCVDeposit.spaceId()).to.be.eq(beforeSnapshotId);
-    });
-
-    it('should be able setDelegate() to give snapshot voting power to another address', async () => {
-      // Can setDelegate() to give Snapshot voting power to someone else
-      expect(await contracts.veBalDelegatorPCVDeposit.delegate()).to.be.equal(contractAddresses.eswak);
-      await vebalOtcHelper.connect(otcBuyerSigner).setDelegate(contractAddresses.feiDAOTimelock);
-      expect(await contracts.veBalDelegatorPCVDeposit.delegate()).to.be.equal(contractAddresses.feiDAOTimelock);
-    });
-
-    it('should be able to clearDelegate() to give snapshot voting power to nobody', async () => {
-      // Can clearDelegate() to give Snapshot voting power to nobody
-      const snapshotSpaceId = await contracts.veBalDelegatorPCVDeposit.spaceId();
-      expect(await contracts.veBalDelegatorPCVDeposit.delegate()).to.be.equal(contractAddresses.feiDAOTimelock);
-      expect(
-        await contracts.snapshotDelegateRegistry.delegation(contractAddresses.veBalDelegatorPCVDeposit, snapshotSpaceId)
-      ).to.be.equal(contractAddresses.feiDAOTimelock);
-      await vebalOtcHelper.connect(otcBuyerSigner).clearDelegate();
-      expect(await contracts.veBalDelegatorPCVDeposit.delegate()).to.be.equal(vebalOtcHelper.address);
-      expect(
-        await contracts.snapshotDelegateRegistry.delegation(contractAddresses.veBalDelegatorPCVDeposit, snapshotSpaceId)
-      ).to.be.equal(ethers.constants.AddressZero);
-    });
-
-    it('can setBalancerMinter()', async () => {
-      const newMinter = contractAddresses.aaveCompaniesMultisig;
-      await vebalOtcHelper.connect(otcBuyerSigner).setBalancerMinter(newMinter);
-      expect(await contracts.balancerGaugeStaker.balancerMinter()).to.equal(newMinter);
-    });
-
-    it('should be able to setGaugeController() to update gauge controller', async () => {
-      await vebalOtcHelper.connect(otcBuyerSigner).setGaugeController(contractAddresses.feiDAOTimelock);
-      expect(await contracts.veBalDelegatorPCVDeposit.gaugeController()).to.equal(contractAddresses.feiDAOTimelock);
-      await vebalOtcHelper.connect(otcBuyerSigner).setGaugeController(contractAddresses.balancerGaugeController);
-      expect(await contracts.veBalDelegatorPCVDeposit.gaugeController()).to.equal(
-        contractAddresses.balancerGaugeController
-      );
-    });
-  });
-
-  describe('Gauge management', () => {
-    it('can stakeInGauge()', async () => {
-      expect(await contracts.balancerGaugeBpt30Fei70Weth.balanceOf(contractAddresses.balancerGaugeStaker)).to.be.equal(
-        0
-      );
-      await contracts.bpt30Fei70Weth
-        .connect(gaugeTokenHolderSigner)
-        .transfer(contractAddresses.balancerGaugeStaker, 100);
-
-      await vebalOtcHelper.connect(otcBuyerSigner).stakeInGauge(contractAddresses.bpt30Fei70Weth, 100);
-
-      expect(await contracts.balancerGaugeBpt30Fei70Weth.balanceOf(contractAddresses.balancerGaugeStaker)).to.be.equal(
-        100
-      );
-    });
-
-    it('can stakeAllInGauge()', async () => {
-      const stakeBeforeBalance = await contracts.balancerGaugeBpt30Fei70Weth.balanceOf(
-        contractAddresses.balancerGaugeStaker
-      );
-      await contracts.bpt30Fei70Weth
-        .connect(gaugeTokenHolderSigner)
-        .transfer(contractAddresses.balancerGaugeStaker, 100);
-
-      await vebalOtcHelper.connect(otcBuyerSigner).stakeAllInGauge(contractAddresses.bpt30Fei70Weth);
-      const stakeBalanceIncrease = (
-        await contracts.balancerGaugeBpt30Fei70Weth.balanceOf(contractAddresses.balancerGaugeStaker)
-      ).sub(stakeBeforeBalance);
-      expect(stakeBalanceIncrease).to.equal(100);
-    });
-
-    it('can unstakeFromGauge()', async () => {
-      const stakeBeforeBalance = await contracts.balancerGaugeBpt30Fei70Weth.balanceOf(
-        contractAddresses.balancerGaugeStaker
-      );
-      await contracts.bpt30Fei70Weth
-        .connect(gaugeTokenHolderSigner)
-        .transfer(contractAddresses.balancerGaugeStaker, 100);
-
-      await vebalOtcHelper.connect(otcBuyerSigner).stakeInGauge(contractAddresses.bpt30Fei70Weth, 100);
-      await vebalOtcHelper.connect(otcBuyerSigner).unstakeFromGauge(contractAddresses.bpt30Fei70Weth, 100);
-      const stakeBalanceDiff = (
-        await contracts.balancerGaugeBpt30Fei70Weth.balanceOf(contractAddresses.balancerGaugeStaker)
-      ).sub(stakeBeforeBalance);
-      expect(stakeBalanceDiff).to.equal(0);
-    });
-
-    it('should be able to voteForGaugeWeight() to vote for gauge weights whilst a lock is active ', async () => {
-      // remove 100% votes for B-30FEI-70WETH
-      expect(
-        (
-          await contracts.balancerGaugeController.vote_user_slopes(
-            contractAddresses.veBalDelegatorPCVDeposit,
-            contractAddresses.balancerGaugeBpt30Fei70Weth
-          )
-        )[1]
-      ).to.be.equal('10000');
-      await vebalOtcHelper
-        .connect(otcBuyerSigner)
-        .voteForGaugeWeight(contractAddresses.bpt30Fei70Weth, contractAddresses.balancerGaugeBpt30Fei70Weth, 0);
-      expect(
-        (
-          await contracts.balancerGaugeController.vote_user_slopes(
-            contractAddresses.veBalDelegatorPCVDeposit,
-            contractAddresses.balancerGaugeBpt30Fei70Weth
-          )
-        )[1]
-      ).to.be.equal('0');
-      // set 100% votes for bb-a-usd
-      await vebalOtcHelper.connect(otcBuyerSigner).voteForGaugeWeight(
-        '0x7B50775383d3D6f0215A8F290f2C9e2eEBBEceb2', // bb-a-usd token
-        '0x68d019f64A7aa97e2D4e7363AEE42251D08124Fb', // bb-a-usd gauge
-        10000
-      );
-      expect(
-        (
-          await contracts.balancerGaugeController.vote_user_slopes(
-            contractAddresses.veBalDelegatorPCVDeposit,
-            '0x68d019f64A7aa97e2D4e7363AEE42251D08124Fb'
-          )
-        )[1]
-      ).to.be.equal('10000');
-    });
-  });
-
-  describe('Vote lock management', () => {
-    it('can lock', async () => {
-      // Transfer BPT tokens to the PCV deposit, for veBalHelper to lock into veBAL
-      await contracts.balWethBPT
-        .connect(balWethBPTWhaleSigner)
-        .transfer(contractAddresses.veBalDelegatorPCVDeposit, 100);
-
-      // Expected to fail, Balancer DAO prevented this veBAL from being relocked
-      await expect(vebalOtcHelper.connect(otcBuyerSigner).lock()).to.be.revertedWith(
-        'Smart contract depositors not allowed'
-      );
-    });
-
-    it('should be able to setLockDuration()', async () => {
-      await contracts.vebalOtcHelper.connect(otcBuyerSigner).setLockDuration(20);
-      expect(await contracts.veBalDelegatorPCVDeposit.lockDuration()).to.be.eq(20);
-    });
-
-    it('should be able to exitLock()', async () => {
-      const veBalBefore = await contracts.veBal.balanceOf(contracts.veBalDelegatorPCVDeposit.address);
-      const balWethBefore = await contracts.balWethBPT.balanceOf(contracts.veBalDelegatorPCVDeposit.address);
-      expect(veBalBefore).to.be.not.eq(0);
-
-      await time.increase(3600 * 24 * 365); // fast foward to end of locking period
-      await vebalOtcHelper.connect(otcBuyerSigner).exitLock();
-
-      const veBalAfter = await contracts.veBal.balanceOf(contracts.veBalDelegatorPCVDeposit.address);
-      const balWethAfter = await contracts.balWethBPT.balanceOf(contracts.veBalDelegatorPCVDeposit.address);
-      expect(veBalAfter).to.be.eq(0);
-      expect(balWethAfter).to.be.gt(balWethBefore);
-    });
-  });
-
-  describe('Boost management', () => {
-    it('should be able to create_boost() to boost delegation to another address', async () => {
-      await vebalOtcHelper.connect(otcBuyerSigner).create_boost(
-        contractAddresses.veBalDelegatorPCVDeposit, // address _delegator
-        contractAddresses.eswak, // address _receiver
-        '10000', // int256 _percentage
-        '1669852800', // uint256 _cancel_time = December 1 2022
-        '1672272000', // uint256 _expire_time = December 29 2022
-        '0' // uint256 _id
-      );
-      const expectedMinBoost = '70000000000000000000000'; // should be 77.5k with 18 decimals as of 14/09/2022
-      expect(
-        await contracts.balancerVotingEscrowDelegation.delegated_boost(contracts.veBalDelegatorPCVDeposit.address)
-      ).to.be.at.least(expectedMinBoost);
-      expect(await contracts.balancerVotingEscrowDelegation.received_boost(contractAddresses.eswak)).to.be.at.least(
-        expectedMinBoost
-      );
-
-      // token id is uint256(delegatorAddress << 96 + boostId), and boostId = 0
-      const tokenId = '0xc4eac760c2c631ee0b064e39888b89158ff808b2000000000000000000000000';
-      expect(await contracts.balancerVotingEscrowDelegation.token_boost(tokenId)).to.be.at.least(expectedMinBoost);
-      expect(await contracts.balancerVotingEscrowDelegation.token_expiry(tokenId)).to.equal('1672272000');
-      expect(await contracts.balancerVotingEscrowDelegation.token_cancel_time(tokenId)).to.equal('1669852800');
-    });
-
-    it('should be able to extend_boost', async () => {
-      // Create boost
-      await vebalOtcHelper.connect(otcBuyerSigner).create_boost(
-        contractAddresses.veBalDelegatorPCVDeposit, // address _delegator
-        contractAddresses.eswak, // address _receiver
-        '10000', // int256 _percentage
-        '1669852800', // uint256 _cancel_time = December 1 2022
-        '1672272000', // uint256 _expire_time = December 29 2022
-        '0' // uint256 _id
-      );
-      const tokenId = '0xc4eac760c2c631ee0b064e39888b89158ff808b2000000000000000000000000';
-
-      await time.increase(86400 * 8);
-      const boostedExpireTime = '1674998582'; // uint256 _expire_time = Jan 29 2023
-      //  - boostedExpireTime gets rounded down to nearest week
-      const boostedCancelTime = '1672694348'; // uint256 _cancel_time = Jan 2 2023
-
-      // Extend the boost
-      await vebalOtcHelper.connect(otcBuyerSigner).extend_boost(
-        tokenId,
-        '10000', // int256 _percentage
-        boostedExpireTime,
-        boostedCancelTime
-      );
-
-      const recordedBoostedExpireTime = await contracts.balancerVotingEscrowDelegation.token_expiry(tokenId);
-
-      // Expected boosted expire time rounded down to nearest week
-      const WEEK = 86400 * 7;
-      const expectedBoostedExpire = Math.floor(Number(boostedExpireTime) / WEEK) * WEEK;
-      expect(recordedBoostedExpireTime).to.equal(expectedBoostedExpire);
-      expect(await contracts.balancerVotingEscrowDelegation.token_cancel_time(tokenId)).to.equal(boostedCancelTime);
-    });
-
-    it('should be able to cancel_boost', async () => {
-      // Create boost
-      await vebalOtcHelper.connect(otcBuyerSigner).create_boost(
-        contractAddresses.veBalDelegatorPCVDeposit, // address _delegator
-        contractAddresses.eswak, // address _receiver
-        '10000', // int256 _percentage
-        '1665432310', // uint256 _cancel_time = October 10 2022
-        '1676495110', // uint256 _expire_time = February 15 2023
-        '0' // uint256 _id
-      );
-      const tokenId = '0xc4eac760c2c631ee0b064e39888b89158ff808b2000000000000000000000000';
-      expect(await contracts.balancerVotingEscrowDelegation.token_boost(tokenId)).to.be.at.least('1'); // non-zero check
-
-      // Cancel boost - fast forward past cancel time
-      await time.increase(86400 * 30);
-      await vebalOtcHelper.connect(otcBuyerSigner).cancel_boost(tokenId);
-
-      const balancerTokenBoost = await contracts.balancerVotingEscrowDelegation.token_boost(tokenId);
-      expect(balancerTokenBoost).to.equal(0);
-    });
-
-    it('should be able to burn a token', async () => {
-      // What is the difference between delegator and receiver
-      // Create boost
-      await vebalOtcHelper.connect(otcBuyerSigner).create_boost(
-        contractAddresses.veBalDelegatorPCVDeposit, // address _delegator
-        contractAddresses.veBalDelegatorPCVDeposit, // address _receiver (becomes owner, only owner can burn)
-        '10000', // int256 _percentage
-        '1669852800', // uint256 _cancel_time = December 1 2022
-        '1672272000', // uint256 _expire_time = December 29 2022
-        '0' // uint256 _id
-      );
-      const tokenId = '0xc4eac760c2c631ee0b064e39888b89158ff808b2000000000000000000000000';
-      // Burn token
-      await vebalOtcHelper.connect(otcBuyerSigner).burn(tokenId);
-    });
-
-    it('should be able to setVotingEscrowDelegation()', async () => {
-      const newDelegation = contractAddresses.dai;
-      await vebalOtcHelper.connect(otcBuyerSigner).setVotingEscrowDelegation(newDelegation);
-      expect(await contracts.balancerGaugeStaker.votingEscrowDelegation()).to.equal(newDelegation);
-    });
-
-    it('should be able to transferBalancerStakerOwnership()', async () => {
-      const newOwner = contractAddresses.dai;
-      await vebalOtcHelper.connect(otcBuyerSigner).transferBalancerStakerOwnership(newOwner);
-      expect(await contracts.balancerGaugeStaker.owner()).to.equal(newOwner);
-
-      // Transfer back
-      const newOwnerSigner = await getImpersonatedSigner(newOwner);
-      await vebalOtcHelper.connect(newOwnerSigner).transferBalancerStakerOwnership(otcBuyerAddress);
-    });
-  });
-
-  describe('Asset management', () => {
-    // Tests withdrawERC20() from pcvDeposit
-    it('should be able to withdrawERC20() to receive B-80BAL-20WETH at end of lock', async () => {
-      await time.increase(3600 * 24 * 365);
-      await vebalOtcHelper.connect(otcBuyerSigner).exitLock();
-
-      // Can withdrawERC20() to receive B-80BAL-20WETH at end of lock
-      const bpt80Bal20WethAmount = await contracts.bpt80Bal20Weth.balanceOf(contractAddresses.veBalDelegatorPCVDeposit);
-      await vebalOtcHelper
-        .connect(otcBuyerSigner)
-        .withdrawERC20(contractAddresses.bpt80Bal20Weth, otcBuyerAddress, bpt80Bal20WethAmount);
-      expect(await contracts.bpt80Bal20Weth.balanceOf(otcBuyerAddress)).to.be.equal(bpt80Bal20WethAmount);
-      expect(bpt80Bal20WethAmount).to.be.at.least(e18(112_041));
-    });
-
-    // Tests withdrawERC20() from staker
-    it('can withdrawERC20 from staker', async () => {
-      await forceEth(balWethBPTWhale);
-
-      // Transfer 1 balWETHBPT token to the balancerGaugeStaker, which we will later withdraw
-      // via the vebalOtcHelper contract
-      await contracts.balWethBPT.connect(balWethBPTWhaleSigner).transfer(contracts.balancerGaugeStaker.address, 1);
-      await forceEth(contracts.balancerGaugeStaker.address);
-
-      const userBalanceBefore = await contracts.balWethBPT.balanceOf(otcBuyerAddress);
-
-      // Withdraw the 1 BPT token from the balancerGaugeStaker to the user
-      await contracts.vebalOtcHelper
-        .connect(otcBuyerSigner)
-        .withdrawERC20fromStaker(contracts.balWethBPT.address, otcBuyerAddress, 1);
-
-      const userBalanceAfter = await contracts.balWethBPT.balanceOf(otcBuyerAddress);
-      expect(userBalanceAfter).to.be.eq(userBalanceBefore.add(1));
-    });
-
-    it('can withdrawETH from pcvDeposit', async () => {
-      // Transfer 1 ETH to the veBalPCVDeposit
-      await forceEth(contractAddresses.veBalDelegatorPCVDeposit);
-      const oneEth = ethers.utils.parseEther('1');
-
-      const receiverAddress = '0x0000000000000000000000000000000000000001';
-      const userEthBefore = await ethers.provider.getBalance(receiverAddress);
-
-      // Withdraw ETH from pcvDeposit to the user
-      await vebalOtcHelper.connect(otcBuyerSigner).withdrawETH(receiverAddress, oneEth);
-
-      const userEthAfter = await ethers.provider.getBalance(receiverAddress);
-      expect(userEthAfter.sub(userEthBefore)).to.equal(oneEth);
-    });
-
-    it('can withdrawETH from staker', async () => {
-      // Transfer 1 ETH to the staker
-      await forceEth(contractAddresses.balancerGaugeStaker);
-      const oneEth = ethers.utils.parseEther('1');
-
-      const receiverAddress = '0x0000000000000000000000000000000000000001';
-      const userEthBefore = await ethers.provider.getBalance(receiverAddress);
-
-      // Withdraw ETH from pcvDeposit to the user
-      await vebalOtcHelper.connect(otcBuyerSigner).withdrawETHfromStaker(receiverAddress, oneEth);
-
-      const userEthAfter = await ethers.provider.getBalance(receiverAddress);
-      expect(userEthAfter.sub(userEthBefore)).to.equal(oneEth);
-    });
-  });
-});
